requirements.txt

annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.30
langchain-groq==0.1.9
langsmith==0.1.99
mariadb==1.1.10
mysqlclient==2.2.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2

annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.30
langchain-groq==0.1.9
langsmith==0.1.99
mariadb==1.1.10
mysqlclient==2.2.4
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
gunicorn
